refactor: restructure folder and files
