+++ ts/inputs/InputNumberOfHexagons.ts
